Move UpgradeTimeFormatToRFC3339 to helper package
