+++ linode/helper/stateupgrade/rfc3339.go
@@ -1,4 +1,4 @@
-package stackscript
+package stateupgrade
 
 import (
 	"time"
@@ -18,3 +18,4 @@ func UpgradeTimeFormatToRFC3339(oldTime string) (timetypes.RFC3339, error) {
 
 	return timetypes.NewRFC3339TimeValue(newTime), nil
 }
+
